requirements.txt
fastapi==0.115.12
telnyx==2.1.5
python-dotenv==1.1.0
uvicorn==0.34.1
